refactor: 코드 개선 및 poll create JWT 수정
