+++ Backend/README.md
@@ -47,3 +47,23 @@
 - 응답 반환 → 처리된 응답을 클라이언트에게 JSON 형식으로 반환
 - 일관된 응답 형식 제공 → 모든 라우트에 일관된 응답 형식을 제공하여 API의 일관성과 유지보수성 향상
     - 프론트가 너무 좋아해요
+
+## 체인을 추가한다 ?
+
+ * Express.js의 미들웨어 또는 라우트 핸들러를 순차적으로 연결하는 개념
+ * Express.js에서 여러 미들웨어나 라우트 핸들러를 하나의 라우트에 연결할 수 있고,
+ * 연결된 것들은 정의된 순서대로 실행되며 이러한 방식이 체이닝(Chaining)이다.
+ *
+ * 예제로 아래와 같이 있다고 가정한다면
+ * router.get('/somepath', handler1, handler2, handler3);
+ * handler1, handler2, handler3dms '/somepath' 경로에 대한 GET 요청을 처리하기 위해 체인으로 연결
+ * 이렇게 되면 Express는 요청이 들어오면 handler1을 먼저 실행하고
+ * handler1이 next()를 호출하면 handler2가 실행되며 순차적으로 실행되는 형식
+ *
+ * 체이닝 방식으로 코드의 모둘화와 재사용성을 높일 수 있고,
+ * 복잡한 라우트 로직을 간단 명확하게 구성을 가능.
+ * 특정 라우트에 대해 필요한 전처리나 후처리를 편리하게 그낭함.
+ *
+ * 결론:
+ * commonResponse 미들웨어를 체인으로 추가하는 경우,
+ * commonResponse는 라우트의 마지막 단계에서 실행되어 일관된 응답 형식으로 반환하게
